REFACTOR: Move notebook cell editors to own submodule.
