+++ packages/frontend/src/notebook/index.ts
@@ -1,2 +1,3 @@
 export * from "./types";
+export * from "./notebook_cells";
 export * from "./notebook_editor";
